+++ OFZ-notification/src/main/java/org/ofz/controller/NotificationController.java
@@ -1,6 +1,9 @@
-package org.ofz;
+package org.ofz.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.ofz.service.NotificationService;
+import org.ofz.dto.SaveTokenRequest;
+import org.ofz.dto.SaveUserTokenResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
